Allow decimal ton quantities in inventory

+++ backend/scripts/migrate-to-neon.js
@@ -47,7 +47,7 @@ async function migrate() {
       id TEXT PRIMARY KEY, name_en TEXT NOT NULL, name_ar TEXT,
       description_en TEXT, description_ar TEXT, price REAL NOT NULL, old_price REAL,
       material TEXT, water_resistance TEXT, size_info TEXT, category_id TEXT,
-      images TEXT DEFAULT '[]', main_image TEXT, stock INTEGER DEFAULT 0,
+      images TEXT DEFAULT '[]', main_image TEXT, stock REAL DEFAULT 0,
       is_active INTEGER DEFAULT 1, is_featured INTEGER DEFAULT 0,
       created_at TIMESTAMPTZ DEFAULT NOW(), updated_at TIMESTAMPTZ DEFAULT NOW()
     );
